% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{subtract}
\alias{subtract}
\title{Function to Subtract}
\usage{
subtract(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{It's another number.}
}
\description{
This function is going to subtract
}
\examples{
subtract (2, 3)

on the console returns [-1]
}
