% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{sum}
\alias{sum}
\title{Function to Sum}
\usage{
sum(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{It's another number.}
}
\description{
This function is going to sum
}
\examples{
sum (2, 3)

on the console returns [5]
}
